refactor(analytics-control-bar): extract date range code into space time utilities
